+++ apps/rush-lib/src/cli/actions/PublishAction.ts
@@ -173,6 +173,7 @@ export class PublishAction extends BaseRushAction {
    * Executes the publish action, which will read change request files, apply changes to package.jsons,
    */
   protected run(): Promise<void> {
+    this.runRushCheckIfNecessary();
     return Promise.resolve().then(() => {
       PolicyValidator.validatePolicy(this.rushConfiguration, false);
